MiniProject_Devops/requirements.txt
Django==5.1
gunicorn==20.0.0
djangorestframework